Move docker env checks into DockerEnvironmentAwareTestContainer
