+++ test/fixtures/minio-fixture/src/main/java/org/elasticsearch/test/fixtures/minio/MinioTestContainer.java
@@ -10,7 +10,6 @@ package org.elasticsearch.test.fixtures.minio;
 
 import org.elasticsearch.test.fixtures.testcontainers.DockerEnvironmentAwareTestContainer;
 import org.junit.rules.TestRule;
-import org.testcontainers.DockerClientFactory;
 import org.testcontainers.images.builder.ImageFromDockerfile;
 
 public class MinioTestContainer extends DockerEnvironmentAwareTestContainer implements TestRule {
@@ -18,21 +17,8 @@ public class MinioTestContainer extends DockerEnvironmentAwareTestContainer impl
     private static final int servicePort = 9000;
     private final boolean enabled;
 
-    /**
-     * see <a href="https://github.com/elastic/elasticsearch/issues/102532">https://github.com/elastic/elasticsearch/issues/102532</a>
-     * */
-    private static boolean isDockerAvailable() {
-        try {
-            DockerClientFactory.instance().client();
-            return true;
-        } catch (Throwable ex) {
-            LOGGER.warn("Probing docker has failed; disabling test", ex);
-            return false;
-        }
-    }
-
     public MinioTestContainer() {
-        this(true && isDockerAvailable());
+        this(true);
     }
 
     public MinioTestContainer(boolean enabled) {
